requirements.txt
# Core dependencies
fastapi
uvicorn
python-dotenv
pydantic
requests

# Additional dependencies
python-multipart